+++ CHANGELOG.md
@@ -8,6 +8,14 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - Representation of chromsomes from GRCh38
 
+## [2.7.19]
+### Added
+- Flag to retain chr/CHR/Chr prefixes when they are present
+
+## [2.7.18]
+### Added
+- `ignore-gq-if-unset` flag to ignore GQ threshold check when GQ or QUAL field is unset for some variants in a VCF file.
+
 ## [2.7.17]
 ### Added
 - Flag to skip GQ check on SV files